+++ packages/sdk-metrics/test/export/PeriodicExportingMetricReader.test.ts
@@ -38,8 +38,9 @@ const MAX_32_BIT_INT = 2 ** 31 - 1;
 class TestMetricExporter implements PushMetricExporter {
   public exportTime = 0;
   public forceFlushTime = 0;
-  public throwException = false;
-  public failureResult = false;
+  public throwExport = false;
+  public throwFlush = false;
+  public rejectExport = false;
   private _batches: ResourceMetrics[] = [];
   private _shutdown: boolean = false;
 
@@ -49,11 +50,11 @@ class TestMetricExporter implements PushMetricExporter {
   ): void {
     this._batches.push(metrics);
 
-    if (this.throwException) {
+    if (this.throwExport) {
       throw new Error('Error during export');
     }
     setTimeout(() => {
-      if (this.failureResult) {
+      if (this.rejectExport) {
         resultCallback({
           code: ExportResultCode.FAILED,
           error: new Error('some error'),
@@ -72,7 +73,7 @@ class TestMetricExporter implements PushMetricExporter {
   }
 
   async forceFlush(): Promise<void> {
-    if (this.throwException) {
+    if (this.throwFlush) {
       throw new Error('Error during forceFlush');
     }
 
@@ -90,6 +91,10 @@ class TestMetricExporter implements PushMetricExporter {
       await new Promise(resolve => setTimeout(resolve, 20));
     }
     return this._batches.slice(0, numberOfExports);
+  }
+
+  getExports(): ResourceMetrics[] {
+    return this._batches.slice(0);
   }
 }
 
@@ -203,7 +208,7 @@ describe('PeriodicExportingMetricReader', () => {
   describe('periodic export', () => {
     it('should keep running on export errors', async () => {
       const exporter = new TestMetricExporter();
-      exporter.throwException = true;
+      exporter.throwExport = true;
       const reader = new PeriodicExportingMetricReader({
         exporter: exporter,
         exportIntervalMillis: 30,
@@ -218,13 +223,13 @@ describe('PeriodicExportingMetricReader', () => {
         emptyResourceMetrics,
       ]);
 
-      exporter.throwException = false;
+      exporter.throwExport = false;
       await reader.shutdown();
     });
 
     it('should keep running on export failure', async () => {
       const exporter = new TestMetricExporter();
-      exporter.failureResult = true;
+      exporter.rejectExport = true;
       const reader = new PeriodicExportingMetricReader({
         exporter: exporter,
         exportIntervalMillis: 30,
@@ -239,7 +244,7 @@ describe('PeriodicExportingMetricReader', () => {
         emptyResourceMetrics,
       ]);
 
-      exporter.failureResult = false;
+      exporter.rejectExport = false;
       await reader.shutdown();
     });
 
@@ -261,7 +266,7 @@ describe('PeriodicExportingMetricReader', () => {
         emptyResourceMetrics,
       ]);
 
-      exporter.throwException = false;
+      exporter.throwExport = false;
       await reader.shutdown();
     });
   });
@@ -271,7 +276,7 @@ describe('PeriodicExportingMetricReader', () => {
       sinon.restore();
     });
 
-    it('should forceFlush exporter', async () => {
+    it('should collect and forceFlush exporter', async () => {
       const exporter = new TestMetricExporter();
       const exporterMock = sinon.mock(exporter);
       exporterMock.expects('forceFlush').calledOnceWithExactly();
@@ -284,6 +289,10 @@ describe('PeriodicExportingMetricReader', () => {
       reader.setMetricProducer(new TestMetricProducer());
       await reader.forceFlush();
       exporterMock.verify();
+
+      const exports = exporter.getExports();
+      assert.strictEqual(exports.length, 1);
+
       await reader.shutdown();
     });
 
@@ -307,12 +316,13 @@ describe('PeriodicExportingMetricReader', () => {
 
     it('should throw when exporter throws', async () => {
       const exporter = new TestMetricExporter();
-      exporter.throwException = true;
+      exporter.throwFlush = true;
       const reader = new PeriodicExportingMetricReader({
         exporter: exporter,
         exportIntervalMillis: MAX_32_BIT_INT,
         exportTimeoutMillis: 80,
       });
+      reader.setMetricProducer(new TestMetricProducer());
 
       await assertRejects(() => reader.forceFlush(), /Error during forceFlush/);
     });
@@ -454,7 +464,7 @@ describe('PeriodicExportingMetricReader', () => {
 
     it('should throw on non-initialized instance.', async () => {
       const exporter = new TestMetricExporter();
-      exporter.throwException = true;
+      exporter.throwFlush = true;
       const reader = new PeriodicExportingMetricReader({
         exporter: exporter,
         exportIntervalMillis: MAX_32_BIT_INT,